requirements.txt
slack_bolt>=1.22.0
fastapi>=0.110.0
starlette>=0.37
uvicorn[standard]>=0.29.0
google-adk>=1.12.0
httpx>=0.27.0
python-dotenv>=1.0.0
aiohttp>=3.9
pytz